Ignore values outside of canvas
Previously any point was written to its internal data structure, but not
to the ppm files.  Now points outside the canvas bounds or not stored
at all.

Asking for a point out side the bounds should return black rather than
cause an error this may change in future to use an optional.